# Add `@inline` for `Diagonal`'s 2-arg `l/rdiv!` to enable auto vectorization