On master, `Diagnal`'s 2-arg `l/rdiv!` call their 3-arg version for code reuse.
But as shown in #43153, non-inlined code blocks LLVM's auto vectorization, so the 2-arg `l/rdiv!` should be slower than before.
This PR first add `@inline` at the call side to fix the regression. But it turns out that `broadcast` based `ldiv!` still blocks vectorization. So I have to replace it with a for loop version.
Some benchmark:
```julia
julia> A = randn(128, 128); D = Diagonal(ones(128));
julia> @btime rdiv!($A, $D);
  7.275 μs (0 allocations: 0 bytes) # about 14 μs on master
julia> @btime ldiv!($D, $A);
  7.350 μs (0 allocations: 0 bytes) # about 14 μs on master
```