@dkarrasch 2021-11-21
I'm not sure it makes a difference, but I haven't seen `@inline` being written on the rhs of function definitions. Don't we typically write this at the beginning of the line?