@N5N3 2021-11-21
See 1.8 release not:
"`@inline` and `@noinline` annotations can now be applied to a function callsite or block to enforce the involved function calls to be (or not to be) inlined. (#41312)"
So `ldiv!(D::Diagonal, B::AbstractVecOrMat) = @inline ldiv!(B, D, B)` could make 3-arg `ldiv!` not be force-inlined by the compiler.