@dkarrasch 2021-11-21
> as we just want to inline it in 2-arg version

Yes, that's why I'm used to seeing the pattern
```julia
@inline ldiv!(D::Diagonal, B::AbstractVecOrMat) = ldiv!(B, D, B)
```
In the one-line case, this may not make a difference. I think what you're referring to makes a difference in multi-line code:
```julia
function foo(...)
    # some code involving possibly non-inlined function calls
    @inline call_a_function(...) # inline that specific function
    # some other code, possibly non-inlined function calls
    return result
end
```