@N5N3 2021-11-21
I'm a little confused by the example in #41312, but macroexpand shows that:
```julia
julia> @macroexpand ldiv!(D::Diagonal, B::AbstractVecOrMat) = @inline ldiv!(B, D, B)
:(ldiv!(D::Diagonal, B::AbstractVecOrMat) = begin
          #= REPL[2]:1 =#
          begin
              $(Expr(:inline, true))
              local var"#37#val" = ldiv!(B, D, B)
              $(Expr(:inline, false))
              var"#37#val"
          end
      end)

julia> @macroexpand @inline ldiv!(D::Diagonal, B::AbstractVecOrMat) = ldiv!(B, D, B)
:(ldiv!(D::Diagonal, B::AbstractVecOrMat) = begin
          $(Expr(:meta, :inline))
          #= REPL[3]:1 =#
          ldiv!(B, D, B)
      end)
julia> @macroexpand ldiv!(D::Diagonal, B::AbstractVecOrMat) = begin
                 @inline
                 ldiv!(B, D, B)
           end
:(ldiv!(D::Diagonal, B::AbstractVecOrMat) = begin
          #= REPL[9]:1 =#
          #= REPL[9]:2 =#
          $(Expr(:meta, :inline))
          #= REPL[9]:3 =#
          ldiv!(B, D, B)
      end)
```
So `ldiv!(D::Diagonal, B::AbstractVecOrMat) = @inline ldiv!(B, D, B)` only tells the compiler to inline the 3-arg `ldiv!` in the 2-arg one. And no inline hint is set for 2-arg version?