@dkarrasch 2021-11-22
Interestingly, on current nightly I don't see the regression for the out-of-place operation:
```julia
julia> using LinearAlgebra, BenchmarkTools

julia> A = randn(128, 128); D = Diagonal(ones(128));

julia> @btime rdiv!($A, $D);
  18.315 μs (0 allocations: 0 bytes)

julia> @btime ldiv!($D, $A);
  18.437 μs (0 allocations: 0 bytes)

julia> @btime $D \ $A;
  9.939 μs (2 allocations: 128.05 KiB)

julia> @btime $A / $D;
  9.614 μs (2 allocations: 128.05 KiB)
```
I'm on MacOS.