@N5N3 2021-11-22
Yes, the regression is on `rdiv!(A, D)`, `ldiv!(D, A)`.
Your bench also shows that they are 2-times slower than `A/D` and `D\A`, which call 3-arg.